Tag Reading works!
Can now read tags and extract the payload message from them. This still
needs to be refined a little as there is something else at the beginning
of the message for some reason. Also added some more audio prompts. Next
up: clean up read payload and send request for it to backend!